[Fix] Keep SGLang sparse MLA indexer in SGLang plugin
