+++ atom/plugin/sglang/models/deepseek_mla.py
@@ -10,6 +10,7 @@ SGLang plugin mode. The heavy DeepSeek-specific runtime helpers live in
 
 from __future__ import annotations
 
+from types import MethodType
 from typing import TYPE_CHECKING, Any
 
 from atom.plugin.sglang.models.deepseek_mla_attention import (
@@ -59,11 +60,54 @@ def _patch_mla_attention_for_sglang(
     _align_qknorm_fusion_for_sglang(attn)
     init_sgl_attrs(attn, config, kv_cache_dtype)
     _patch_attention_projs_for_sglang_mxfp4(attn)
+    _patch_indexer_for_sglang_sparse_mla(attn)
     if not isinstance(attn.mla_attn, SGLangDeepseekMLAAttention):
         attn.mla_attn = SGLangDeepseekMLAAttention(attn, attn.mla_attn)
     attn.process_weights_after_loading = lambda: process_mla_kv_b_proj_after_loading(
         attn
     )
+
+
+def _patch_indexer_for_sglang_sparse_mla(attn: "DeepseekV2MLAAttention") -> None:
+    """Adapt DeepSeek-V3.2 sparse indexer buffers for SGLang plugin mode."""
+    indexer = getattr(attn, "indexer", None)
+    if indexer is None or getattr(indexer, "_atom_sglang_topk_buffer_patched", False):
+        return
+
+    import torch
+    import atom.plugin.sglang.attention_backend.sparse_mla_indexer  # noqa: F401
+
+    original_forward = indexer.forward
+    indexer.use_qk_rope_cache_fusion = False
+    indexer.sparse_attn_indexer_impl = (
+        torch.ops.aiter.sparse_attn_indexer_sglang_plugin_mode
+    )
+
+    def _forward_with_topk_buffer(self, hidden_states, *args, **kwargs):
+        num_tokens = int(hidden_states.shape[0])
+        topk_tokens = int(self.topk_tokens)
+        buffer = getattr(self, "topk_indices_buffer", None)
+        needs_new_buffer = (
+            buffer is None
+            or buffer.dim() != 2
+            or buffer.device != hidden_states.device
+            or buffer.shape[0] < num_tokens
+            or buffer.shape[1] < topk_tokens
+        )
+        if needs_new_buffer:
+            buffer = torch.empty(
+                num_tokens,
+                topk_tokens,
+                dtype=torch.int32,
+                device=hidden_states.device,
+            )
+            self.topk_indices_buffer = buffer
+
+        self.sparse_kv_indices_buffer = self.topk_indices_buffer
+        return original_forward(hidden_states, *args, **kwargs)
+
+    indexer.forward = MethodType(_forward_with_topk_buffer, indexer)
+    indexer._atom_sglang_topk_buffer_patched = True
 
 
 def _align_qknorm_fusion_for_sglang(attn: "DeepseekV2MLAAttention") -> None:
