+++ atom/plugin/vllm/attention/layer_sparse_mla.py
@@ -20,12 +20,11 @@ from aiter import (
     indexer_k_quant_and_cache,
     indexer_qk_rope_quant_and_cache,
     top_k_per_row_decode,
-    top_k_per_row_prefill,
 )
 from aiter.ops.triton.fp8_mqa_logits import fp8_mqa_logits
 from aiter.ops.triton.pa_mqa_logits import deepgemm_fp8_paged_mqa_logits
 
-from atom.plugin.prepare import is_sglang, is_vllm
+from atom.plugin.prepare import is_vllm
 from atom.utils.custom_register import direct_register_custom_op
 
 import triton
@@ -33,7 +32,6 @@ import triton.language as tl
 
 from typing import Optional
 import logging
-import re
 
 logger = logging.getLogger("atom")
 
@@ -506,250 +504,6 @@ def sparse_attn_indexer_fake(
     return torch.empty(weights.shape, device=weights.device, dtype=torch.float32)
 
 
-def _parse_layer_id_from_indexer_prefix(prefix: str) -> int:
-    match = re.search(r"\.layers\.(\d+)\.", prefix)
-    if match is None:
-        raise RuntimeError(
-            f"Cannot infer DeepSeek-V3.2 indexer layer id from prefix: {prefix!r}"
-        )
-    return int(match.group(1))
-
-
-def _build_sglang_query_ranges(forward_batch) -> tuple[torch.Tensor, torch.Tensor]:
-    device = forward_batch.seq_lens.device
-    if forward_batch.forward_mode.is_decode_or_idle():
-        bs = int(forward_batch.batch_size)
-        starts = torch.zeros(bs, dtype=torch.int32, device=device)
-        ends = forward_batch.seq_lens[:bs].to(dtype=torch.int32)
-        return starts, ends
-
-    query_lens = getattr(forward_batch, "extend_seq_lens", None)
-    if query_lens is None:
-        query_lens = forward_batch.seq_lens
-    query_lens_cpu = getattr(forward_batch, "extend_seq_lens_cpu", None)
-    if query_lens_cpu is None:
-        query_lens_cpu = query_lens.detach().cpu()
-    seq_lens_cpu = getattr(forward_batch, "seq_lens_cpu", None)
-    if seq_lens_cpu is None:
-        seq_lens_cpu = forward_batch.seq_lens.detach().cpu()
-
-    starts = []
-    ends = []
-    kv_offset = 0
-    for q_len_raw, seq_len_raw in zip(query_lens_cpu, seq_lens_cpu):
-        q_len = int(q_len_raw)
-        seq_len = int(seq_len_raw)
-        prefix_len = seq_len - q_len
-        starts.extend([kv_offset] * q_len)
-        ends.extend(kv_offset + prefix_len + i + 1 for i in range(q_len))
-        kv_offset += seq_len
-
-    return (
-        torch.tensor(starts, dtype=torch.int32, device=device),
-        torch.tensor(ends, dtype=torch.int32, device=device),
-    )
-
-
-def _build_sglang_block_table(forward_batch, page_size: int) -> torch.Tensor:
-    req_pool_indices = forward_batch.req_pool_indices
-    req_to_token = forward_batch.req_to_token_pool.req_to_token
-    token_table = req_to_token[req_pool_indices, :]
-    if not forward_batch.forward_mode.is_decode_or_idle():
-        token_table = token_table.clone()
-        query_lens = getattr(forward_batch, "extend_seq_lens", None)
-        if query_lens is None:
-            query_lens = forward_batch.seq_lens
-        prefix_lens = getattr(forward_batch, "extend_prefix_lens", None)
-        if prefix_lens is None:
-            prefix_lens = forward_batch.seq_lens - query_lens
-        query_lens_cpu = query_lens[: int(forward_batch.batch_size)].detach().cpu()
-        prefix_lens_cpu = prefix_lens[: int(forward_batch.batch_size)].detach().cpu()
-        offset = 0
-        for req_idx, (prefix_len_raw, query_len_raw) in enumerate(
-            zip(prefix_lens_cpu, query_lens_cpu)
-        ):
-            prefix_len = int(prefix_len_raw)
-            query_len = int(query_len_raw)
-            if query_len > 0:
-                token_table[
-                    req_idx, prefix_len : prefix_len + query_len
-                ] = forward_batch.out_cache_loc[offset : offset + query_len]
-            offset += query_len
-    if page_size == 1:
-        return token_table
-    return token_table[:, ::page_size] // page_size
-
-
-def sparse_attn_indexer_sglang_plugin_mode(
-    hidden_states: torch.Tensor,
-    k_cache_prefix: str,
-    kv_cache: torch.Tensor,
-    q_input: torch.Tensor,
-    k: torch.Tensor,
-    weights: torch.Tensor,
-    quant_block_size: int,
-    scale_fmt: Optional[str],
-    topk_tokens: int,
-    head_dim: int,
-    max_model_len: int,
-    total_seq_lens: int,
-    topk_indices_buffer: torch.Tensor,
-    k_norm_weight: torch.Tensor,
-    k_norm_bias: torch.Tensor,
-    k_norm_eps: float,
-    positions: torch.Tensor,
-    cos_cache: torch.Tensor,
-    sin_cache: torch.Tensor,
-    weights_scale: float,
-    is_neox_style: bool,
-    use_qk_rope_cache_fusion: bool,
-) -> torch.Tensor:
-    from atom.plugin.sglang.models.base_model_wrapper import get_current_forward_batch
-
-    del kv_cache
-    forward_batch = get_current_forward_batch()
-    if forward_batch is None or forward_batch.forward_mode.is_idle():
-        return torch.zeros_like(weights, dtype=torch.float32)
-
-    token_to_kv_pool = forward_batch.token_to_kv_pool
-    if not hasattr(token_to_kv_pool, "get_index_k_with_scale_buffer"):
-        raise RuntimeError(
-            "[SGL+ATOM] DeepSeek-V3.2 sparse MLA requires SGLang NSA KV pool "
-            "with index_k_with_scale_buffer support."
-        )
-
-    layer_id = _parse_layer_id_from_indexer_prefix(k_cache_prefix)
-    index_cache = token_to_kv_pool.get_index_k_with_scale_buffer(layer_id)
-    page_size = int(getattr(token_to_kv_pool, "page_size", 1))
-    kv_cache = index_cache.view(-1, page_size, head_dim + 4)
-    preshuffle_cache = page_size != 1
-    slot_mapping = forward_batch.out_cache_loc
-
-    if use_qk_rope_cache_fusion:
-        q_bf16 = q_input
-        q_fp8 = torch.empty_like(q_bf16, dtype=dtypes.fp8)
-        weights_out = torch.empty(
-            weights.shape, device=weights.device, dtype=torch.float32
-        )
-        indexer_qk_rope_quant_and_cache(
-            q_bf16,
-            q_fp8,
-            weights,
-            weights_out,
-            k,
-            kv_cache,
-            slot_mapping,
-            k_norm_weight,
-            k_norm_bias,
-            positions,
-            cos_cache,
-            sin_cache,
-            k_norm_eps,
-            quant_block_size,
-            scale_fmt,
-            weights_scale,
-            preshuffle=preshuffle_cache,
-            is_neox=is_neox_style,
-        )
-        weights = weights_out
-    else:
-        q_fp8 = q_input
-        indexer_k_quant_and_cache(
-            k,
-            kv_cache,
-            slot_mapping,
-            quant_block_size,
-            scale_fmt,
-            preshuffle=preshuffle_cache,
-        )
-
-    num_tokens = hidden_states.shape[0]
-    topk_indices_buffer[:num_tokens] = -1
-    block_table = _build_sglang_block_table(forward_batch, page_size)
-
-    if forward_batch.forward_mode.is_decode_or_idle():
-        bs = int(forward_batch.batch_size)
-        if q_fp8.shape[0] < bs or weights.shape[0] < bs:
-            raise RuntimeError(
-                "[SGL+ATOM] sparse indexer decode expected at least "
-                f"{bs} token rows, got q={q_fp8.shape[0]}, weights={weights.shape[0]}. "
-                "This usually means TP-scattered indexer inputs were not gathered."
-            )
-        seq_lens_i32 = forward_batch.seq_lens[:bs].to(dtype=torch.int32)
-        padded_q_fp8 = q_fp8[:bs].reshape(bs, 1, *q_fp8.shape[1:])
-        logits = torch.empty([bs, max_model_len], dtype=torch.float32, device=k.device)
-        deepgemm_fp8_paged_mqa_logits(
-            padded_q_fp8,
-            kv_cache.unsqueeze(-2),
-            weights[:bs],
-            logits,
-            seq_lens_i32,
-            block_table,
-            max_model_len,
-            ChunkK=256,
-            Preshuffle=preshuffle_cache,
-            KVBlockSize=page_size,
-            WavePerEU=2,
-        )
-        top_k_per_row_decode(
-            logits,
-            1,
-            seq_lens_i32,
-            topk_indices_buffer[:bs, :topk_tokens],
-            bs,
-            logits.stride(0),
-            logits.stride(1),
-        )
-        return weights
-
-    cu_starts, cu_ends = _build_sglang_query_ranges(forward_batch)
-    total_kv = int(forward_batch.seq_lens_sum)
-    k_fp8 = torch.empty([total_kv, head_dim], device=k.device, dtype=dtypes.fp8)
-    k_scale = torch.empty([total_kv, 1], device=k.device, dtype=torch.float32)
-    cp_gather_indexer_k_quant_cache(
-        kv_cache,
-        k_fp8,
-        k_scale.view(dtypes.fp8),
-        block_table,
-        torch.nn.functional.pad(
-            torch.cumsum(forward_batch.seq_lens, dim=0, dtype=torch.int32), (1, 0)
-        ),
-        preshuffle=preshuffle_cache,
-    )
-    logits = fp8_mqa_logits(
-        Q=q_fp8[:num_tokens],
-        KV=k_fp8,
-        kv_scales=k_scale,
-        weights=weights[:num_tokens],
-        cu_starts=cu_starts,
-        cu_ends=cu_ends,
-    )
-    assert topk_tokens == 2048, "top_k_per_row assumes size 2048"
-    topk_indices = topk_indices_buffer[:num_tokens, :topk_tokens]
-    top_k_per_row_prefill(
-        logits=logits,
-        rowStarts=cu_starts,
-        rowEnds=cu_ends,
-        indices=topk_indices,
-        values=None,
-        numRows=logits.shape[0],
-        stride0=logits.stride(0),
-        stride1=logits.stride(1),
-    )
-    topk_indices.copy_(
-        torch.where(topk_indices >= 0, topk_indices - cu_starts[:, None], topk_indices)
-    )
-    return weights
-
-
-direct_register_custom_op(
-    op_name="sparse_attn_indexer_sglang_plugin_mode",
-    op_func=sparse_attn_indexer_sglang_plugin_mode,
-    mutates_args=["topk_indices_buffer"],
-    fake_impl=sparse_attn_indexer_fake,
-)
-
-
 direct_register_custom_op(
     op_name="sparse_attn_indexer_plugin_mode",
     op_func=sparse_attn_indexer_plugin_mode,
@@ -769,14 +523,6 @@ def IndexerDecoratorForPluginMode(cls):
         if is_vllm():
             self.sparse_attn_indexer_impl = (
                 torch.ops.aiter.sparse_attn_indexer_plugin_mode
-            )
-        elif is_sglang():
-            # SGLang's NSA KV pool uses its own indexer cache layout on HIP.
-            # The fused qk-rope-cache kernel currently aborts before Python can
-            # report a useful error, so use the decomposed SGLang indexer path.
-            self.use_qk_rope_cache_fusion = False
-            self.sparse_attn_indexer_impl = (
-                torch.ops.aiter.sparse_attn_indexer_sglang_plugin_mode
             )
 
     cls.__init__ = new_init
